장바구니 도서 삭제, 주문 (상세) 내역 조회에 로그인 확인

+++ controller/cartController.js
@@ -51,7 +51,7 @@ const getCartItems = (req, res) => {
      where user_id = ?`;
         let values = [authorization.id];
 
-        if(selected){ // 주문서 작성 시 '선택한 장바구니 목록 조회'
+        if (selected) { // 주문서 작성 시 '선택한 장바구니 목록 조회'
             sql += ' and cartItems.id in (?)';
             values.push(selected);
         }
@@ -68,17 +68,30 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id;
 
-    let sql = 'delete from cartItems where id = ?';
-    conn.query(sql, cartItemId, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    let authorization = ensureAuthorization(req, res);
 
-        return res.status(StatusCodes.OK).json(results);
-    });
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message': '로그인 세션이 만료되었습니다. 다시 로그인하세요.'
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            'message': '잘못된 토큰입니다.'
+        });
+    } else {
+        const cartItemId = req.params.id;
+
+        let sql = 'delete from cartItems where id = ?';
+        conn.query(sql, cartItemId, (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+
+            return res.status(StatusCodes.OK).json(results);
+        });
+    }
 };
 
 module.exports = {
